Consolidate DOM price formatting
