+++ storefronts/core/currency/webflow-dom.js
@@ -1,0 +1,1 @@
+export * from '../../platforms/webflow/webflow-dom.js';
